authorization implemented

+++ src/services/notes/controller/note-controller.js
@@ -1,3 +1,4 @@
+import AuthorizationError from '../../../exceptions/authorization-error.js';
 import InvariantError from '../../../exceptions/invariant-error.js';
 import NotFoundError from '../../../exceptions/not-found-error.js';
 import response from '../../../utils/response.js';
@@ -5,10 +6,12 @@ import NoteRepositories from '../repositories/note-repositories.js';
 
 export const createNote = async (req, res, next) => {
   const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
   const note = await NoteRepositories.createNote({
     title,
     body,
     tags,
+    owner,
   });
 
   if (!note) {
@@ -21,12 +24,23 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
   return response(res, 200, 'Catatan sukses ditampilkan', { notes: notes });
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
+
   const note = await NoteRepositories.getNoteById(id);
 
   if (!note) {
@@ -41,6 +55,17 @@ export const getNoteById = async (req, res, next) => {
 export const editNoteById = async (req, res, next) => {
   const { id } = req.params;
   const { title, body, tags } = req.validated;
+
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
+
   const note = await NoteRepositories.editNote({
     id,
     title,
@@ -51,11 +76,22 @@ export const editNoteById = async (req, res, next) => {
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
   }
+
   return response(res, 200, 'Catatan berhasil diperbarui', note);
 };
 
 export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError('Anda tidak berhak mengakses resource ini'),
+    );
+  }
+
   const deletedNote = await NoteRepositories.deleteNote(id);
 
   if (!deletedNote) {
